Split ReflectionUtils
